chore(integrations): move httplib,httpx,jinja2,kafka,kombu to internal
