+++ ddtrace/contrib/httplib/__init__.py
@@ -58,9 +58,10 @@ The integration can be configured per instance::
 
 :ref:`Headers tracing <http-headers-tracing>` is supported for this integration.
 """
-from .patch import get_version
-from .patch import patch
-from .patch import unpatch
+
+from ..internal.httplib.patch import get_version
+from ..internal.httplib.patch import patch
+from ..internal.httplib.patch import unpatch
 
 
 __all__ = ["patch", "unpatch", "get_version"]
