+++ ddtrace/contrib/httpx/__init__.py
@@ -84,8 +84,12 @@ required_modules = ["httpx"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
-        from .patch import unpatch
+        # Required to allow users to import from `ddtrace.contrib.httpx.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.httpx.patch import get_version
+        from ..internal.httpx.patch import patch
+        from ..internal.httpx.patch import unpatch
 
         __all__ = ["patch", "unpatch", "get_version"]
