+++ ddtrace/contrib/internal/jinja2/constants.py
@@ -1,0 +1,1 @@
+DEFAULT_TEMPLATE_NAME = "<memory>"
