+++ ddtrace/contrib/internal/kombu/constants.py
@@ -1,0 +1,1 @@
+DEFAULT_SERVICE = "kombu"
