+++ ddtrace/contrib/jinja2/__init__.py
@@ -34,8 +34,12 @@ required_modules = ["jinja2"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
-        from .patch import unpatch
+        # Required to allow users to import from `ddtrace.contrib.jinja2.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.jinja2.patch import get_version
+        from ..internal.jinja2.patch import patch
+        from ..internal.jinja2.patch import unpatch
 
         __all__ = ["patch", "unpatch", "get_version"]
