+++ ddtrace/contrib/kafka/__init__.py
@@ -48,8 +48,12 @@ required_modules = ["confluent_kafka"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
-        from .patch import unpatch
+        # Required to allow users to import from `ddtrace.contrib.kafka.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.kafka.patch import get_version
+        from ..internal.kafka.patch import patch
+        from ..internal.kafka.patch import unpatch
 
         __all__ = ["patch", "unpatch", "get_version"]
