+++ ddtrace/contrib/kombu/__init__.py
@@ -39,7 +39,11 @@ required_modules = ["kombu", "kombu.messaging"]
 
 with require_modules(required_modules) as missing_modules:
     if not missing_modules:
-        from .patch import get_version
-        from .patch import patch
+        # Required to allow users to import from `ddtrace.contrib.kombu.patch` directly
+        from . import patch as _  # noqa: F401, I001
+
+        # Expose public methods
+        from ..internal.kombu.patch import get_version
+        from ..internal.kombu.patch import patch
 
         __all__ = ["patch", "get_version"]
